move component folder
