+++ src/app/(app)/page.client.tsx
@@ -1,7 +1,5 @@
 "use client";
 
-import { InvoiceClientPage } from "@/app/components";
-import { InvoicePDFDownloadLink } from "@/app/components/invoice-pdf-download-link";
 import { INITIAL_INVOICE_DATA } from "@/app/constants";
 import {
   invoiceSchema,
@@ -38,6 +36,8 @@ import { useEffect, useState } from "react";
 import { toast } from "sonner";
 import { z } from "zod";
 import { VIDEO_DEMO_URL } from "@/config";
+import { InvoicePDFDownloadLink } from "./components/invoice-pdf-download-link";
+import { InvoiceClientPage } from "./components";
 // import { InvoicePDFDownloadMultipleLanguages } from "./components/invoice-pdf-download-multiple-languages";
 
 /**
